Refactoring backend logic and Modify file structure
