+++ src/public/images/favicon/android-icon-192x192.png
