+++ src/public/images/favicon/apple-icon-precomposed.png
