+++ src/public/images/favicon/browserconfig.xml
