+++ src/public/images/favicon/manifest.json
